Do not register CP_COMMAND more than once
Firmata caches sysex response handles in a static context, and throws an error if you attempt to register a handler that's already been registered.  Because of this, it's currently impossible to instatiate Playground more than once. I interpret this as a mistake, since Firmata itself _can_ be instantiated more than once.  The tests didn't catch this because they stub the `sysexResponse` method.

Uses a static flag to make sure we only register the CP_COMMAND sysex response once, even if we instantiate multiple Playground instances.